Data.Propagator

Documentation

data Cell s a

A Cell contains "information about a value" rather than a value per se.

data Change a

This represents the sorts of changes we can make as we accumulate information in a Cell.

Change False indicates that this is the old value, and we didn't change anything.

Change True indicates that this is the new value, which gains information over the old.

Contradiction indicates that the updated information is inconsistent with the old.

Constructors

Change !Bool a

Contradiction String

class Propagated a where

This class provides the default definition for how to merge values in our information lattice.

Minimal complete definition

Nothing

Methods

merge :: a -> a -> Change a

class Propagated a => PropagatedNum a

class PropagatedNum a => PropagatedFloating a

class (Floating a, Ord a) => PropagatedInterval a

Minimal complete definition

infinity

cell :: Propagated a => ST s (Cell s a)

Construct a new Cell with no information.

cellWith :: (a -> a -> Change a) -> ST s (Cell s a)

Construct a new Cell with a custom merge strategy.

known :: Propagated a => a -> ST s (Cell s a)

Construct a Cell with some information

write :: Cell s a -> a -> ST s ()

Writing to a Cell tells it information about its value.

require :: Cell s a -> ST s (Maybe a)

Requiring the value of a Cell will return the value of the cell if present.

with :: Cell s a -> (a -> ST s ()) -> ST s ()

with will read the current value of a Cell and do something with that result if it is known. If the Cell is currently empty, this will do nothing. Unlike watch it does not install a handler.

watch :: Cell s a -> (a -> ST s ()) -> ST s ()

Watching a Cell sets up a callback that will be notified if the cell changes.

watch2 :: Cell s a -> Cell s b -> (a -> b -> ST s ()) -> ST s ()

watch2 will watch two cells. When they both have some information the supplied callback will fire at least once. It will continue to fire each time they get more information from then out.

lift1 :: (a -> b) -> Cell s a -> Cell s b -> ST s ()

Lift a unary function into a relationship between two cells.

lift2 :: (a -> b -> c) -> Cell s a -> Cell s b -> Cell s c -> ST s ()

Lift a binary function into a relationship between two cells.

data Prop s a where

This type allows us to write seemingly normal functional code and glue it together out of smaller propagator templates. Evaluation of these expressions uses Observable Sharing.

Nullary lifts a computation that will produce a Cell into a Prop.

Unary lifts a relationship between 2 cells into Prop.

Binary lifts a relationship between 3 cells into Prop.

Constructors

Nullary :: ST s (Cell s a) -> Prop s a

Unary :: Propagated b => (Cell s a -> Cell s b -> ST s ()) -> Prop s a -> Prop s b

Binary :: Propagated c => (Cell s a -> Cell s b -> Cell s c -> ST s ()) -> Prop s a -> Prop s b -> Prop s c

lower :: Prop s a -> ST s (Cell s a)

Lower a Prop to its output Cell by observable sharing.

arg :: Cell s a -> Prop s a

Lift a Cell into a Prop

lower1 :: (Prop s a -> Prop s b) -> Cell s a -> ST s (Cell s b)

Lower a unary Prop computation to a relationship between two cells.

lower2 :: (Prop s a -> Prop s b -> Prop s c) -> Cell s a -> Cell s b -> ST s (Cell s c)

Lower a binary Prop computation to a relationship between three cells.

forwards :: (Propagated a, Propagated b) => (forall s. Prop s a -> Prop s b) -> a -> Maybe b

Run a Prop formula forwards.

>>> forwards (\c -> c * 9/5 + 32) 100
Just 212.0

backwards :: (Propagated a, Propagated b) => (forall s. Prop s a -> Prop s b) -> b -> Maybe a

Run a Prop formula backwards.

>>> backwards (\c -> c * 9/5 + 32) 212
Just 100.0