Data.Propagator.Cell

Documentation

data Cell s a

A Cell contains "information about a value" rather than a value per se.

Constructors

Cell (a -> a -> Change a) !(MutVar s (Maybe a, a -> ST s ()))

cell :: Propagated a => ST s (Cell s a)

Construct a new Cell with no information.

cellWith :: (a -> a -> Change a) -> ST s (Cell s a)

Construct a new Cell with a custom merge strategy.

known :: Propagated a => a -> ST s (Cell s a)

Construct a Cell with some information

write :: Cell s a -> a -> ST s ()

Writing to a Cell tells it information about its value.

unify :: Cell s a -> Cell s a -> ST s ()

Unifying two cells makes them exchange information as if they were one Cell.

require :: Cell s a -> ST s (Maybe a)

Requiring the value of a Cell will return the value of the cell if present.

watch :: Cell s a -> (a -> ST s ()) -> ST s ()

Watching a Cell sets up a callback that will be notified if the cell changes.

with :: Cell s a -> (a -> ST s ()) -> ST s ()

with will read the current value of a Cell and do something with that result if it is known. If the Cell is currently empty, this will do nothing. Unlike watch it does not install a handler.

watch2 :: Cell s a -> Cell s b -> (a -> b -> ST s ()) -> ST s ()

watch2 will watch two cells. When they both have some information the supplied callback will fire at least once. It will continue to fire each time they get more information from then out.

lift1 :: (a -> b) -> Cell s a -> Cell s b -> ST s ()

Lift a unary function into a relationship between two cells.

lift2 :: (a -> b -> c) -> Cell s a -> Cell s b -> Cell s c -> ST s ()

Lift a binary function into a relationship between two cells.