Data.Propagator.Class

Documentation

data Change a

This represents the sorts of changes we can make as we accumulate information in a Cell.

Change False indicates that this is the old value, and we didn't change anything.

Change True indicates that this is the new value, which gains information over the old.

Contradiction indicates that the updated information is inconsistent with the old.

Constructors

Change !Bool a

Contradiction String

class Propagated a where

This class provides the default definition for how to merge values in our information lattice.

Minimal complete definition

Nothing

Methods

merge :: a -> a -> Change a

mergeDefault :: (Eq a, Show a) => a -> a -> Change a

This is a viable default definition for merge for most simple values.