Data.Propagator.Prop

Documentation

data Prop s a where

This type allows us to write seemingly normal functional code and glue it together out of smaller propagator templates. Evaluation of these expressions uses Observable Sharing.

Nullary lifts a computation that will produce a Cell into a Prop.

Unary lifts a relationship between 2 cells into Prop.

Binary lifts a relationship between 3 cells into Prop.

Constructors

Nullary :: ST s (Cell s a) -> Prop s a

Unary :: Propagated b => (Cell s a -> Cell s b -> ST s ()) -> Prop s a -> Prop s b

Binary :: Propagated c => (Cell s a -> Cell s b -> Cell s c -> ST s ()) -> Prop s a -> Prop s b -> Prop s c

lower :: Prop s a -> ST s (Cell s a)

Lower a Prop to its output Cell by observable sharing.

arg :: Cell s a -> Prop s a

Lift a Cell into a Prop

lower1 :: (Prop s a -> Prop s b) -> Cell s a -> ST s (Cell s b)

Lower a unary Prop computation to a relationship between two cells.

lower2 :: (Prop s a -> Prop s b -> Prop s c) -> Cell s a -> Cell s b -> ST s (Cell s c)

Lower a binary Prop computation to a relationship between three cells.

forwards :: (Propagated a, Propagated b) => (forall s. Prop s a -> Prop s b) -> a -> Maybe b

Run a Prop formula forwards.

>>> forwards (\c -> c * 9/5 + 32) 100
Just 212.0

backwards :: (Propagated a, Propagated b) => (forall s. Prop s a -> Prop s b) -> b -> Maybe a

Run a Prop formula backwards.

>>> backwards (\c -> c * 9/5 + 32) 212
Just 100.0